pyproject.toml
[project]
name = "voicescreen"
version = "0.1.0"
description = "AI Voice Screening Agent for mass hiring"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110,<1",
    "uvicorn[standard]>=0.29,<1",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.29,<1",
    "alembic>=1.13,<2",
    "celery[redis]>=5.3,<6",
    "redis>=5.0,<6",
    "httpx>=0.27,<1",
    "pydantic>=2.6,<3",
    "pydantic-settings>=2.2,<3",
    "structlog>=24.1,<25",
    "tenacity>=8.2,<9",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0,<7",
    "openai>=1.14,<2",
    "aiogram>=3.4,<4",
    "boto3>=1.34,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.23,<1",
    "ruff>=0.3,<1",
    "black>=24.2,<25",
    "mypy>=1.9,<2",
    "pre-commit>=3.6,<4",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.black]
target-version = ["py311"]
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends._legacy:_Backend"
